Moved Incubator ChipsInput to components, fixed review notes
