+++ src/components/chipsInput/assets/xSmall@3x.png
