+++ src/components/chipsInput/assets/xSmall@4x.png
